Add comparator function to sort EC points
Its purpose is to make points on lower-right side (x < 0 and y > 0) as first on the list
and upper-right points as the last ones
(iterating the list of EC points is like walking on the EC curve from lower-right side to upper-right).
When discriminant parameter (4*a^3 + 27*b^2) is < 0 then graph has 2 separated parts and y values less than 0 are the first one on the sorted list.

This function passed unit-tests of sorting for 3 scenarios of discriminant parameter (> 0, = 0 and < 0):
 sorted list was compared with expected one by "assert-array-equals" test.